pyproject.toml
[project]
name = "tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "RodionSOK",email = "socoljv_ev@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "npm (>=0.1.1,<0.2.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
